pyproject.toml
[project]
name = "test-user-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.25",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]
